trading-bot/monitoring/requirements.txt
# Monitoring System Dependencies
fastapi==0.68.2
uvicorn==0.15.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==0.19.2
pydantic==1.10.13
prometheus-client==0.12.0
psutil==5.9.5
dash==2.14.2
dash-bootstrap-components==1.4.1
plotly==5.18.0
pandas==1.5.3
numpy==1.26.4
requests==2.31.0
python-dateutil==2.8.2
Jinja2==3.0.3  # Pinned for Starlette 0.14.2 compatibility

# Security
aiohttp==3.9.1  # For async HTTP requests
python-json-logger==2.0.7  # Structured logging
python-keycloak==3.4.0  # OAuth2/OpenID Connect
python-multipart==0.0.6  # File uploads
python-slugify==8.0.1  # URL-safe strings
python-ulid==2.3.0  # Secure unique IDs
pyyaml==6.0.1  # Safe YAML parsing

# Rate limiting and security headers
slowapi==0.1.8  # Rate limiting
python-jose[cryptography]==3.3.0  # JWT support
python-multipart==0.0.6  # Form parsing
python-slugify==8.0.1  # URL-safe strings

# Monitoring and observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.16.0
opentelemetry-instrumentation-fastapi==0.39b0
opentelemetry-exporter-otlp==1.16.0
opentelemetry-instrumentation-logging==0.39b0
opentelemetry-instrumentation-requests==0.39b0

# Development
types-python-dateutil==2.8.19.14
types-requests==2.31.0.10
types-pyyaml==6.0.12.20250326
